// src/interfaces/QuestionInterface.tsx
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface QuestionData {
    response_code: number;
    results:       Question[];
}

export interface Question {
    category:          string;
    type:              string;
    difficulty:        string;
    question:          string;
    correct_answer:    string;
    incorrect_answers: string[];
    result:            string; 
}